requirements.txt
chembl_webresource_client
rdkit-pypi
typing
nglview
pathlib